backend/test/requirements.txt
# ChainlessChain 后端接口测试依赖

# HTTP 请求
requests>=2.31.0

# 环境变量管理
python-dotenv>=1.0.0

# 终端颜色输出
colorama>=0.4.6

# JSON 处理（Python 内置，无需安装）
# json

# 时间和日期处理（Python 内置，无需安装）
# datetime
